% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{base_64_to_char}
\alias{base_64_to_char}
\title{read json base64 contents from github}
\usage{
base_64_to_char(b64)
}
\arguments{
\item{b64}{base64 character value to be decoded and converted to character value}
}
\description{
provide tool to read and process data using the github api
}
