% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{$.tt_data}
\alias{$.tt_data}
\title{Access TidyTuesday data from tt_data object}
\usage{
\method{$}{tt_data}(x, i)
}
\arguments{
\item{x}{data frame.}

\item{i}{elements to extract or replace.  For \code{[} and
    \code{[[}, these are \code{numeric} or \code{character} or, for
    \code{[} only, empty.  Numeric values are coerced to integer as if
    by \code{\link{as.integer}}.  For replacement by \code{[}, a logical
    matrix is allowed.}
}
\description{
Access TidyTuesday data from tt_data object
}
