% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-conditions.R
\name{.pkg_abort}
\alias{.pkg_abort}
\title{Raise a package-scoped error}
\usage{
.pkg_abort(
  message,
  subclass,
  call = rlang::caller_env(),
  message_env = rlang::caller_env(),
  ...
)
}
\arguments{
\item{message}{(\code{character}) The message for the new error. Messages will be
formatted with \code{\link[cli:cli_bullets]{cli::cli_bullets()}}.}

\item{subclass}{(\code{character}) Class(es) to assign to the error. Will be
prefixed by "\{package\}-error-".}

\item{call}{\code{(environment)} The execution environment to mention as the
source of error messages.}

\item{message_env}{(\code{environment}) The execution environment to use to
evaluate variables in error messages.}

\item{...}{Additional parameters passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} and on to
\code{\link[rlang:abort]{rlang::abort()}}.}
}
\value{
Does not return.
}
\description{
Raise a package-scoped error
}
\keyword{internal}
