% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{gh_get}
\alias{gh_get}
\title{Get data from the tt github repo.}
\usage{
gh_get(path, auth = gh::gh_token(), ...)
}
\arguments{
\item{path}{Path within the rfordatascience/tidytuesday repo.}

\item{auth}{A GitHub token see [gh::gh_token()] for more details.}

\item{...}{Additional parameters passed to [gh::gh()].}
}
\value{
The GitHub response as parsed by [gh::gh()].
}
\description{
Get data from the tt github repo.
}
\keyword{internal}
