% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{github_blob}
\alias{github_blob}
\title{Read blob Contents from GitHub}
\usage{
github_blob(path, as_raw = FALSE, auth = github_pat())
}
\arguments{
\item{path}{Relative path from within the TidyTuesday Repository to contents,
usually because it was too large to be read with the contencts api.}

\item{auth}{github PAT. See PAT section for more information}

\item{raw}{optional arguments to pass to \code{read_html}}
}
\value{
a raw/character object based on the blob
}
\description{
provide tools to read and process blob's using the github api
}
\section{PAT}{


A Github PAT is a personal Access Token. This allows for signed queries to
the github api, and increases the limit on the number of requests allowed from
60 to 5000. Follow instructions https://happygitwithr.com/github-pat.html
to set the PAT.
}

\examples{
\dontrun{

main_readme <- github_html("README.md")
week_readme <- github_html("data/2020/2020-01-07/readme.md")

}

}
