% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{github_contents}
\alias{github_contents}
\title{Read Contents from GitHub}
\usage{
github_contents(path, auth = github_pat())
}
\arguments{
\item{path}{Relative path from within the TidyTuesday Repository}

\item{auth}{github PAT. See PAT section for more information}
}
\value{
raw text of the content with the sha as an attribute
}
\description{
Provide tool to read raw data and return as text the raw data using the github api
}
\section{PAT}{


A Github PAT is a personal Access Token. This allows for signed queries to
the github api, and increases the limit on the number of requests allowed from
60 to 5000. Follow instructions https://happygitwithr.com/github-pat.html
to set the PAT.
}

\examples{
\dontrun{
text_csv <- github_contents("data/2020/2020-04-07/tdf_stages.csv")
tour_de_france_stages <- read_csv(text_csv)

}

}
