% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{github_GET}
\alias{github_GET}
\title{Get for github API}
\usage{
github_GET(url, auth = github_pat(), ...)
}
\arguments{
\item{url}{URL to GET from}

\item{auth}{github PAT}

\item{...}{any additional headers to add}
}
\value{
response from GET
}
\description{
Extract the github PAT from the system environment for authenticated requests.
}
