% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{github_html}
\alias{github_html}
\title{Read Contents from GitHub as html}
\usage{
github_html(path, ...)
}
\arguments{
\item{path}{Relative path from within the TidyTuesday Repository to contents that can be returned as HTML}

\item{...}{optional arguments to pass to \code{read_html}}
}
\value{
result of read_html on the contents
}
\description{
provide tools to read and process readme's as html using the github api
}
\examples{
\dontrun{

main_readme <- github_html("README.md")
week_readme <- github_html("data/2020/2020-01-07/readme.md")

}

}
