% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{github_page}
\alias{github_page}
\title{Create shell for HTML content from github}
\usage{
github_page(page_content)
}
\arguments{
\item{content}{html content}
}
\value{
xml_document with github header
}
\description{
Provide the necessary <head> section to wrap around raw html content read from github.
}
