% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{github_sha}
\alias{github_sha}
\title{Read Contents from GitHub as html}
\usage{
github_sha(dirpath, branch = "master", auth = github_pat())
}
\arguments{
\item{dirpath}{Relative path from within the TidyTuesday Repository to folder of contents wanting sha for}

\item{branch}{which branch to get sha for. assumed to be master (and usually should be)}

\item{auth}{github PAT. See PAT section for more information}
}
\value{
result data.frame of SHA and other information of directory contents
}
\description{
provide tools to read and process readme's as html using the github api
}
\section{PAT}{


A Github PAT is a personal Access Token. This allows for signed queries to
the github api, and increases the limit on the number of requests allowed from
60 to 5000. Follow instructions https://happygitwithr.com/github-pat.html
to set the PAT.
}

\examples{
\dontrun{

sha <- github_sha("data/2020/2020-01-07")

}

}
