% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.tt_data}
\alias{print.tt_data}
\title{print utility for tt_data objects}
\usage{
\method{print}{tt_data}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print utility for tt_data objects
}
