% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_available.R
\name{print.tt_dataset_table}
\alias{print.tt_dataset_table}
\title{print utility for tt_dataset_table object}
\usage{
\method{print}{tt_dataset_table}(x, ..., is_interactive = interactive())
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{is_interactive}{is the console interactive}
}
\description{
print utility for tt_dataset_table object
}
