% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_read_data.R
\name{read_rda}
\alias{read_rda}
\title{utility to load RDA with out using assigned name in envir}
\usage{
read_rda(path)
}
\arguments{
\item{path}{path to RDA file}
}
\description{
utility to load RDA with out using assigned name in envir
}
