% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_record.R
\name{tt_record}
\alias{tt_record}
\alias{tt_playback}
\title{Record and generate plot histories}
\usage{
tt_record(
  dir = NULL,
  device = c("png", "jpeg", "bmp", "tiff", "emf", "svg", "eps"),
  scale = 1,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm"),
  dpi = 300,
  limitsize = TRUE
)

tt_playback(
  name = NULL,
  last_image_duration = 12,
  frame_duration = 0.25,
  ...,
  playback = TRUE,
  stoprecording = TRUE
)
}
\arguments{
\item{dir}{directory to save the intermediate plots in}

\item{device}{Device to use. Can either be a device function
(e.g. \code{\link[=png]{png()}}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{scale}{Multiplicative scaling factor.}

\item{width}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{height}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{units}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{limitsize}{When \code{TRUE} (the default), \code{ggsave} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}

\item{name}{name of gif}

\item{last_image_duration}{n units of frame_duration to show the last image for}

\item{frame_duration}{n seconds each plot should be shown}

\item{...}{arguments passed to \code{\link[gifski]{gifski}}}

\item{playback}{Boolean, should the recording start playing after it is
turned into a gif? defaults to TRUE}

\item{stoprecording}{Boolean, should the plots stop being recorded?
defaults to TRUE.}
}
\description{
Record your past plots and generate a gif of the plots you've
   made in this r session. Common technique used by TidyTuesday-ers to show
   where they started vs their final plots
}
\section{Functions}{
\itemize{
\item \code{tt_playback}: 
}}

\examples{

if(require(ggplot2)){
  tt_record(dir = file.path(tempdir(),"recording"))
  ggplot(data.frame(x = 1, y = 1), aes(x=x, y=y)) + geom_point() + ylim(0,4)
  ggplot(data.frame(x = 1, y = 2), aes(x=x, y=y)) + geom_point() + ylim(0,4)
  ggplot(data.frame(x = 1, y = 3), aes(x=x, y=y)) + geom_point() + ylim(0,4)
  tt_playback(tempfile(fileext = ".gif"))
}

}
