% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_master_file.R
\name{should_update_tt_master_file}
\alias{should_update_tt_master_file}
\title{Decide whether to update the master file}
\usage{
should_update_tt_master_file(force = FALSE, auth = gh::gh_token())
}
\arguments{
\item{force}{force the update to occur even if the SHA matches}
}
\value{
Boolean indicating whether the master file should be updated.
}
\description{
Decide whether to update the master file
}
\keyword{internal}
