% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_readme}
\alias{show_readme}
\title{Print Readme to RStudio HTML Viewer}
\usage{
show_readme(tt)
}
\arguments{
\item{tt}{tt_data object for printing}
}
\description{
Print Readme to RStudio HTML Viewer
}
