% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_check_date.R
\name{tt_check_date}
\alias{tt_check_date}
\title{Generate valid TidyTuesday URL}
\usage{
tt_check_date(x, week = NULL, auth = gh::gh_token())
}
\arguments{
\item{x}{The date of data to pull (in "YYYY-MM-dd" format), or the four-digit
year as a number.}

\item{week}{Which week number to use within a given year. Only used when \code{x}
is a valid year.}

\item{auth}{A GitHub token. See \code{\link[gh:gh_token]{gh::gh_token()}} for more details.}
}
\description{
Given multiple types of inputs, generate a valid TidyTuesday URL.
}
\keyword{internal}
