% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_compile.R
\name{tt_compile}
\alias{tt_compile}
\title{Get TidyTuesday Readme and list of files and HTML}
\usage{
tt_compile(date)
}
\arguments{
\item{date}{date of tidytuesday of interest}
}
\description{
Get TidyTuesday Readme and list of files and HTML
}
