% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_available.R
\name{tt_datasets}
\alias{tt_datasets}
\title{Available datasets}
\usage{
tt_datasets(year, auth = github_pat())
}
\arguments{
\item{year}{numeric entry representing the year of tidytuesday you want the list of datasets
for. Leave empty for most recent year.}

\item{auth}{github Personal Access Token. See PAT section for more information}
}
\description{
list available datasets for that year
}
\section{PAT}{


A Github PAT is a personal Access Token. This allows for signed queries to
the github api, and increases the limit on the number of requests allowed from
60 to 5000. Follow instructions https://happygitwithr.com/github-pat.html
to set the PAT.
}

