% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\docType{data}
\name{TT_GITHUB_ENV}
\alias{TT_GITHUB_ENV}
\title{Environment containing state of Github API limits}
\format{
An object of class \code{environment} of length 3.
}
\usage{
TT_GITHUB_ENV
}
\description{
Environment containing state of Github API limits
}
\keyword{internal}
