% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_make_url.R
\name{tt_make_url}
\alias{tt_make_url}
\title{given inputs generate valid TidyTuesday URL}
\usage{
tt_make_url(x, week)
}
\arguments{
\item{x}{either a string or numeric entry indicating the full date of}
}
\description{
Given multiple types of inputs, generate
}
