% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_master_file.R
\docType{data}
\name{TT_MASTER_ENV}
\alias{TT_MASTER_ENV}
\title{The Master List of Files from TidyTuesday}
\format{
An object of class \code{environment} of length 1.
}
\usage{
TT_MASTER_ENV
}
\description{
The Master List of Files from TidyTuesday
}
\keyword{internal}
