% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_master_file.R
\name{tt_master_file}
\alias{tt_master_file}
\title{Get Master List of Files from TidyTuesday}
\usage{
tt_master_file(force = FALSE, auth = gh::gh_token())
}
\arguments{
\item{force}{force the update to occur even if the SHA matches}

\item{auth}{A GitHub token see [gh::gh_token()] for more details.}
}
\value{
The tt master file, updated if necessary.
}
\description{
Import or update dataset from github that records the entire list of objects
from TidyTuesday
}
\keyword{internal}
