% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_parse.R
\name{tt_parse_binary}
\alias{tt_parse_binary}
\title{utility to assist with parsing the raw binary data}
\usage{
tt_parse_binary(blob, func, ..., filename)
}
\arguments{
\item{blob}{raw data to be parsed}

\item{func}{the function to perform parsing of the file}

\item{...}{args to pass to func}

\item{filename}{the original name of the file}
}
\description{
utility to assist with parsing the raw binary data
}
