% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_parse.R
\name{tt_parse_blob}
\alias{tt_parse_blob}
\title{general utility to assist with parsing the raw data}
\usage{
tt_parse_blob(blob, ..., file_info)
}
\arguments{
\item{blob}{raw data to be parsed}

\item{...}{args to pass to func}

\item{file_info}{data.frame of information about the blob being read}
}
\description{
general utility to assist with parsing the raw data
}
