% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_parse.R
\name{tt_parse_text}
\alias{tt_parse_text}
\title{utility to assist with parsing the text data}
\usage{
tt_parse_text(blob, func, ...)
}
\arguments{
\item{blob}{raw text data to be parsed}

\item{func}{the function to perform parsing of the file}

\item{...}{args to pass to func}
}
\description{
utility to assist with parsing the text data
}
