% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_read_data.R
\name{tt_read_data}
\alias{tt_read_data}
\title{Reads in TidyTuesday datasets from Github repo}
\usage{
tt_read_data(tt, x)
}
\arguments{
\item{tt}{tt_gh object from tt_load_gh function}

\item{x}{index/name of data object to read in. string or int}
}
\value{
tibble
}
\description{
Reads in the actual data from the TidyTuesday github
}
\examples{
tt_gh<-tt_load_gh("2019-01-15")

tt_dataset_1<-tt_read_data(tt_gh,tt_gh$files[1])
}
\concept{tt_read_data}
