% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_master_file.R
\name{tt_update_master_file}
\alias{tt_update_master_file}
\title{Get Master List of Files from TidyTuesday}
\usage{
tt_update_master_file(force = FALSE)
}
\description{
Import or update dataset from github that records the entire list of objects from tidytuesday
}
\keyword{internal}
