% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_read_data.R
\name{download_read}
\alias{download_read}
\title{utility to assist with 'reading' urls that cannot normally be read by file functions}
\usage{
download_read(url, func, ..., guess_max, mode = "w",
  find_delim = FALSE)
}
\arguments{
\item{url}{path to online file to be read}

\item{func}{the function to perform reading of url}

\item{...}{args to pass to func}

\item{mode}{mode passed to \code{utils::download.file}. default is "w"}

\item{find_delim}{should the delimeters be found for the file}
}
\description{
utility to assist with 'reading' urls that cannot normally be read by file functions
}
