% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{gh_get}
\alias{gh_get}
\title{Get data from the tt github repo.}
\usage{
gh_get(path, auth = gh::gh_token(), ...)
}
\arguments{
\item{path}{Path within the \code{rfordatascience/tidytuesday} repo.}

\item{...}{Additional parameters passed to \code{\link[gh:gh]{gh::gh()}}.}
}
\value{
The GitHub response as parsed by \code{\link[gh:gh]{gh::gh()}}.
}
\description{
Get data from the tt github repo.
}
\keyword{internal}
