% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{github_blob}
\alias{github_blob}
\title{Read blob Contents from GitHub}
\usage{
github_blob(path, as_raw = FALSE, sha = NULL, auth = github_pat())
}
\arguments{
\item{path}{Relative path from within the TidyTuesday Repository to contents,
usually because it was too large to be read with the contencts api.}

\item{as_raw}{optional arguments to pass to \code{read_html}}

\item{sha}{sha of object if known in liu of path (usually best to give both for clarity)}

\item{auth}{github PAT. See PAT section for more information}
}
\value{
a raw/character object based on the blob
}
\description{
provide tools to read and process blob's using the github api
}
\section{PAT}{


A Github PAT is a personal Access Token. This allows for signed queries to
the github api, and increases the limit on the number of requests allowed from
60 to 5000. Follow instructions https://happygitwithr.com/github-pat.html
to set the PAT.
}

\examples{
\dontrun{

main_readme_blob <- github_blob("README.md", as_raw = TRUE)

}

}
