% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{github_contents}
\alias{github_contents}
\title{Read Contents from GitHub}
\usage{
github_contents(path, read_func = read.csv, ...)
}
\arguments{
\item{path}{Relative path from within the TidyTuesday Repository}

\item{read_func}{Function to parse the text. Defaults to \code{read.csv}}

\item{...}{optional arguments to pass to \code{read_func}}
}
\value{
result of read_func on the content
}
\description{
Provide tool to read raw data and return as text the raw data using the github api
}
