% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{github_html}
\alias{github_html}
\title{Read Contents from GitHub as html}
\usage{
github_html(path, ..., auth = github_pat())
}
\arguments{
\item{path}{Relative path from within the TidyTuesday Repository to contents that can be returned as HTML}

\item{...}{optional arguments to pass to \code{read_html}}

\item{auth}{github PAT. See PAT section for more information}
}
\value{
result of read_html on the contents
}
\description{
Provide tools to read and process readme's as html using the github api
}
\section{PAT}{


A Github PAT is a personal Access Token. This allows for signed queries to
the github api, and increases the limit on the number of requests allowed from
60 to 5000. Follow instructions https://happygitwithr.com/github-pat.html
to set the PAT.
}

\examples{
\dontrun{

main_readme <- github_html("README.md")
week_readme <- github_html("data/2020/2020-01-07/readme.md")

}

}
