% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_available.R
\name{github_page}
\alias{github_page}
\title{Create shell for HTML content from github}
\usage{
github_page(page_content)
}
\arguments{
\item{page_content}{html content in xml_document class}
}
\value{
xml_document with github header
}
\description{
Provide the necessary <head> section to wrap around raw html content read
from github. This is only actually used locally for printing, and may not be
necessary anymore.
}
\keyword{internal}
