% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{github_pat}
\alias{github_pat}
\title{Get the github PAT}
\usage{
github_pat(quiet = TRUE)
}
\arguments{
\item{quiet}{Should this be loud? default TRUE.}
}
\value{
PAT as a character.
}
\description{
Extract the github PAT from the system environment for authenticated requests.
}
