% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_tuesday.R
\name{last_tuesday}
\alias{last_tuesday}
\title{Find the most recent tuesday}
\usage{
last_tuesday(date = today(tz = "America/New_York"))
}
\arguments{
\item{date}{todays date as a date object}
}
\description{
Utility function to assist users in identifying the most recent 'tidytuesday' date
}
\examples{

last_tuesday() # get last tuesday from todays date
last_tuesday(as.Date("2020-01-01")) # get last tuesday from specified date

}
