% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_tuesday.R
\name{last_tuesday}
\alias{last_tuesday}
\title{Find the most recent tuesday}
\usage{
last_tuesday(date = today(tzone = "America/New_York"))
}
\arguments{
\item{date}{todays date as a date object or character string in
YYYY-MM-DD format.}
}
\description{
Utility function to assist users in identifying the most recent 'TidyTuesday' date
}
\examples{

last_tuesday() # get last tuesday from todays date
last_tuesday("2020-01-01") # get last tuesday from specified date

}
