% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_tuesday.R
\name{last_tuesday}
\alias{last_tuesday}
\title{Find the most recent tuesday}
\usage{
last_tuesday(date = today(tzone = "America/New_York"))
}
\arguments{
\item{date}{A date as a date object or character string in \code{YYYY-MM-DD}
format. Defaults to today's date.}
}
\description{
Identify the most recent 'TidyTuesday' date relative to a specified date.
}
\examples{
last_tuesday() # get last Tuesday relative to today's date
last_tuesday("2020-01-01") # get last Tuesday relative to a specified date

}
