% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_available.R
\name{print.tt_dataset_table_list}
\alias{print.tt_dataset_table_list}
\title{print utility for tt_dataset_table_list object}
\usage{
\method{print}{tt_dataset_table_list}(x, ..., is_interactive = interactive())
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{is_interactive}{is the console interactive}
}
\value{
used for side effects to show all the available datasets in TidyTuesday
}
\description{
print utility for tt_dataset_table_list object
}
\examples{

available_datasets <- tt_available()
print(available_datasets)

}
