% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readme}
\alias{readme}
\title{Readme HTML maker and Viewer}
\usage{
readme(tt)
}
\arguments{
\item{tt}{tt_data object for printing}
}
\value{
Null, invisibly. Used to show readme of the downloaded TidyTuesday
  dataset in the Viewer.
}
\description{
Readme HTML maker and Viewer
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
if (rate_limit_check(quiet = TRUE) > 30) {
  tt_output <- tt_load_gh("2019-01-15")
  readme(tt_output)
}
\dontshow{\}) # examplesIf}
}
