% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readme}
\alias{readme}
\title{Readme HTML maker and Viewer}
\usage{
readme(tt)
}
\arguments{
\item{tt}{tt_data object for printing}
}
\value{
Does not return anything. Used to show readme of the downloaded tidytuesday dataset in the Viewer.
}
\description{
Readme HTML maker and Viewer
}
\examples{
tt_output <- tt_load_gh("2019-01-15")
readme(tt_output)
}
