% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{auth}{A GitHub token. See \code{\link[gh:gh_token]{gh::gh_token()}} for more details.}

\item{tt}{A \code{tt} object, output from \code{\link[=tt_load_gh]{tt_load_gh()}}.}

\item{files}{Which file names to download. Default "All" downloads all files
for the specified week.}

\item{week}{Which week number to use within a given year. Only used when \code{x}
is a valid year.}

\item{x}{The date of data to pull (in "YYYY-MM-dd" format), or the four-digit
year as a number.}

\item{year}{What year of TidyTuesday to use}
}
\description{
Reused parameter definitions are gathered here for easier editing.
}
\keyword{internal}
