% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytuesdayR-package.R
\docType{package}
\name{tidytuesdayR-package}
\alias{tidytuesdayR}
\alias{tidytuesdayR-package}
\title{tidytuesdayR: Access the Weekly 'TidyTuesday' Project Dataset}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

'TidyTuesday' is a project by the 'Data Science Learning Community' in which they post a weekly dataset in a public data repository (\url{https://github.com/rfordatascience/tidytuesday}) for people to analyze and visualize. This package provides the tools to easily download this data and the description of the source.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thebioengineer.github.io/tidytuesdayR/}
  \item \url{https://github.com/thebioengineer/tidytuesdayR}
  \item Report bugs at \url{https://github.com/thebioengineer/tidytuesdayR/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Authors:
\itemize{
  \item Ellis Hughes \email{ellishughes@live.com}
}

Other contributors:
\itemize{
  \item Thomas Mock \email{j.thomasmock@gmail.com} [contributor]
  \item Data Science Learning Community \email{tidytuesday@dslc.io} [data contributor]
}

}
\keyword{internal}
