% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_available.R
\name{tt_available}
\alias{tt_available}
\title{Show all TidyTuesdays}
\usage{
tt_available(auth = github_pat())
}
\arguments{
\item{auth}{github Personal Access Token. See PAT section for more information}
}
\description{
Show all the available datasets, and corresponding weeks
}
\section{PAT}{


A Github PAT is a personal Access Token. This allows for signed queries to
the github api, and increases the limit on the number of requests allowed from
60 to 5000. Follow instructions https://happygitwithr.com/github-pat.html
to set the PAT.
}

