% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_check_date.R
\name{tt_check_date}
\alias{tt_check_date}
\title{Generate valid TidyTuesday URL}
\usage{
tt_check_date(x, week, auth = gh::gh_token())
}
\arguments{
\item{x}{either a string or numeric entry indicating the full date of the
dataset.}

\item{week}{left empty unless x is a numeric year entry, in which case the
week of interest should be entered}

\item{auth}{A GitHub token see [gh::gh_token()] for more details.}
}
\description{
Given multiple types of inputs, generate a valid TidyTuesday URL.
}
\keyword{internal}
