% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_make_url.R
\name{tt_check_date}
\alias{tt_check_date}
\title{given inputs generate valid TidyTuesday URL}
\usage{
tt_check_date(x, week)
}
\arguments{
\item{x}{either a string or numeric entry indicating the full date of}

\item{week}{left empty unless x is a numeric year entry, in which case the week of interest should be entered}
}
\description{
Given multiple types of inputs, generate
}
