% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_curate_data.R
\name{tt_curate_data}
\alias{tt_curate_data}
\title{Guidance for TidyTuesday dataset curation}
\usage{
tt_curate_data()
}
\value{
The path to the \code{tt_curation.R} script, invisibly.
}
\description{
Open an R script to guide you through the process of curating and submitting
a TidyTuesday dataset. See \verb{vignette("curating", package = "tidytuesdayR)}
for more information.
}
\examples{
tt_curate_data()
}
