% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_datasets.R
\name{tt_datasets}
\alias{tt_datasets}
\title{Available datasets}
\usage{
tt_datasets(year)
}
\arguments{
\item{year}{numeric entry representing the year of tidytuesday you want the list of datasets for. Leave empty for most recent year.}
}
\description{
list available datasets for that year
}
