% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_check_date.R
\name{tt_date}
\alias{tt_date}
\title{Get date of Tidytuesday, given the year and week}
\usage{
tt_date(year, week)
}
\arguments{
\item{year}{what year of TidyTuesday to use}

\item{week}{what week of Tidytuesday to use}
}
\description{
Sometimes we don't know the date we want, but we do know the
week. this function provides the ability to pass the year and week we are
interested in to get the correct date
}
\examples{
\donttest{
if(interactive()){
 tt_date(2019, week = 42)
 }
}
}
