% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_check_date.R
\name{tt_date}
\alias{tt_date}
\title{Get date of TidyTuesday, given the year and week}
\usage{
tt_date(year, week, auth = gh::gh_token())
}
\arguments{
\item{year}{What year of TidyTuesday to use}

\item{week}{What week of Tidytuesday to use}

\item{auth}{A GitHub token see [gh::gh_token()] for more details.}
}
\description{
Sometimes we don't know the date we want, but we do know the week. This
function provides the ability to pass the year and week we are interested in
to get the correct date
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tt_date(2019, week = 42)
\dontshow{\}) # examplesIf}
}
