% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_download.R
\name{tt_download}
\alias{tt_download}
\title{download tt data

Download all or specific files identified in the tt dataset}
\usage{
tt_download(tt, files = c("All", attr(tt, ".files")$data_files), ...)
}
\arguments{
\item{tt}{string representation of the date of data to pull, in YYYY-MM-dd format, or just numeric entry for year}

\item{files}{List the file names to download. Default to asking.}

\item{...}{pass methods to the parsing functions. These will be passed to ALL files, so be careful.}
}
\value{
tt_data object (list class)
}
\description{
download tt data

Download all or specific files identified in the tt dataset
}
\examples{
tt_output <- tt_load("2019-01-15")
}
