% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_load.R
\name{tt_load}
\alias{tt_load}
\title{Load tt data from Github}
\usage{
tt_load(date)
}
\arguments{
\item{date}{YYYY-MM-dd date string}
}
\value{
tt_data object (list class)
}
\description{
Load tt data from Github
}
\examples{
tt_output<-tt_load("2019-01-15")


}
