% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_load.R
\name{tt_load_gh}
\alias{tt_load_gh}
\title{Load TidyTuesday data from Github}
\usage{
tt_load_gh(x, week)
}
\arguments{
\item{x}{string representation of the date of data to pull, in YYYY-MM-dd format, or just numeric entry for year}

\item{week}{left empty unless x is a numeric year entry, in which case the week of interest should be entered}
}
\value{
tt_gh object. List object with the following entries: readme, files, url
}
\description{
Pulls the Readme and URLs of the data from the TidyTuesday github folder based on the date provided
}
\examples{
tt_gh <- tt_load_gh("2019-01-15")

readme(tt_gh)
}
