% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_master_file.R
\name{tt_master_file}
\alias{tt_master_file}
\title{Get Master List of Files from Local Environment}
\usage{
tt_master_file(assign = NULL)
}
\arguments{
\item{assign}{value to overwrite the TT_MASTER_ENV$TT_MASTER_FILE contents with}
}
\description{
return or update tt master file
}
\keyword{internal}
