% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_master_file.R
\name{tt_update_master_file}
\alias{tt_update_master_file}
\title{Get Master List of Files from TidyTuesday}
\usage{
tt_update_master_file(force = FALSE, auth = github_pat())
}
\arguments{
\item{force}{force the update to occur even if the SHA matches}

\item{auth}{github Personal Access Token. See PAT section for more information}
}
\description{
Import or update dataset from github that records the entire list of objects from tidytuesday
}
\section{PAT}{


A Github PAT is a personal Access Token. This allows for signed queries to
the github api, and increases the limit on the number of requests allowed from
60 to 5000. Follow instructions https://happygitwithr.com/github-pat.html
to set the PAT.
}

\keyword{internal}
