% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tidytemplate.R
\name{use_tidytemplate}
\alias{use_tidytemplate}
\title{Call and open the tidytemplate}
\usage{
use_tidytemplate(name = "My_Awesome_TidyTuesday_Plot.Rmd")
}
\description{
Use the tidytemplate Rmd for starting your analysis with a leg up for processing
}
