% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tidytemplate.R
\name{use_tidytemplate}
\alias{use_tidytemplate}
\title{Call and open the tidytemplate}
\usage{
use_tidytemplate(name = NULL, open = interactive(), ..., refdate = today())
}
\arguments{
\item{name}{name of your TidyTuesday analysis file}

\item{open}{should the file be opened after being created}

\item{...}{arguments to be passed to \code{\link[usethis:use_template]{usethis::use_template()}}}

\item{refdate}{date to use as reference to determine which TidyTuesday to
use for the template. Either date object or character string in
YYYY-MM-DD format.}
}
\description{
Use the tidytemplate Rmd for starting your analysis with a
leg up for processing
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
use_tidytemplate(name = "My_Awesome_TidyTuesday.Rmd")
\dontshow{\}) # examplesIf}
}
